Move SMPP multipart message handling from protocol to service.
