+++ vumi/transports/smpp/tests/test_smpp_service.py
@@ -8,7 +8,8 @@ from vumi.transports.smpp.smpp_transport import (
     SmppTransceiverTransport, SmppMessageDataStash)
 from vumi.transports.smpp.protocol import EsmeProtocol, EsmeProtocolError
 from vumi.transports.smpp.smpp_service import SmppService
-from vumi.transports.smpp.pdu_utils import command_id
+from vumi.transports.smpp.pdu_utils import (
+    command_id, unpacked_pdu_opts, short_message)
 from vumi.transports.smpp.sequence import RedisSequence
 from vumi.transports.smpp.tests.fake_smsc import FakeSMSC
 
@@ -93,6 +94,10 @@ class TestSmppService(VumiTestCase):
         """
         lookup_func = service.message_stash.get_sequence_number_message_id
         return gatherResults([lookup_func(seq_num) for seq_num in seq_nums])
+
+    def set_sequence_number(self, service, seq_nr):
+        return service.sequence_generator.redis.set(
+            'smpp_last_sequence_number', seq_nr)
 
     @inlineCallbacks
     def test_start_sequence(self):
@@ -204,3 +209,236 @@ class TestSmppService(VumiTestCase):
         unbind_resp_pdu = yield self.fake_smsc.await_pdu()
         self.assertEqual(command_id(unbind_resp_pdu), 'unbind_resp')
         self.assertEqual(service.is_bound(), False)
+
+    @inlineCallbacks
+    def test_csm_split_message(self):
+        """
+        A multipart message is split into chunks such that the smallest number
+        of message parts are required.
+        """
+        service = yield self.get_service()
+
+        split = lambda msg: service.csm_split_message(msg.encode('utf-8'))
+
+        # these are fine because they're in the 7-bit character set
+        self.assertEqual(1, len(split(u'&' * 140)))
+        self.assertEqual(1, len(split(u'&' * 160)))
+        # ± is not in the 7-bit character set so it should utf-8 encode it
+        # which bumps it over the 140 bytes
+        self.assertEqual(2, len(split(u'±' + u'1' * 139)))
+
+    @inlineCallbacks
+    def test_submit_sm_long(self):
+        """
+        A long message can be sent in a single PDU using the optional
+        `message_payload` PDU field.
+        """
+        service = yield self.get_service()
+        yield self.fake_smsc.bind()
+
+        long_message = 'This is a long message.' * 20
+        seq_nums = yield service.submit_sm_long(
+            'abc123', 'dest_addr', long_message)
+        submit_sm = yield self.fake_smsc.await_pdu()
+        pdu_opts = unpacked_pdu_opts(submit_sm)
+
+        self.assertEqual('submit_sm', submit_sm['header']['command_id'])
+        self.assertEqual(
+            None, submit_sm['body']['mandatory_parameters']['short_message'])
+        self.assertEqual(''.join('%02x' % ord(c) for c in long_message),
+                         pdu_opts['message_payload'])
+        stored_ids = yield self.lookup_message_ids(service, seq_nums)
+        self.assertEqual(['abc123'], stored_ids)
+
+    @inlineCallbacks
+    def test_submit_csm_sar(self):
+        """
+        A long message can be sent in multiple PDUs with SAR fields set to
+        instruct the SMSC to build user data headers.
+        """
+        service = yield self.get_service({'send_multipart_sar': True})
+        yield self.fake_smsc.bind()
+
+        long_message = 'This is a long message.' * 20
+        seq_nums = yield service.submit_csm_sar(
+            'abc123', 'dest_addr', short_message=long_message)
+        pdus = yield self.fake_smsc.await_pdus(4)
+        # seq no 1 == bind_transceiver, 2 == enquire_link, 3 == sar_msg_ref_num
+        self.assertEqual([4, 5, 6, 7], seq_nums)
+        msg_parts = []
+        msg_refs = []
+
+        for i, sm in enumerate(pdus):
+            pdu_opts = unpacked_pdu_opts(sm)
+            mandatory_parameters = sm['body']['mandatory_parameters']
+
+            self.assertEqual('submit_sm', sm['header']['command_id'])
+            msg_parts.append(mandatory_parameters['short_message'])
+            self.assertTrue(len(mandatory_parameters['short_message']) <= 130)
+            msg_refs.append(pdu_opts['sar_msg_ref_num'])
+            self.assertEqual(i + 1, pdu_opts['sar_segment_seqnum'])
+            self.assertEqual(4, pdu_opts['sar_total_segments'])
+
+        self.assertEqual(long_message, ''.join(msg_parts))
+        self.assertEqual([3, 3, 3, 3], msg_refs)
+
+        stored_ids = yield self.lookup_message_ids(service, seq_nums)
+        self.assertEqual(['abc123'] * len(seq_nums), stored_ids)
+
+    @inlineCallbacks
+    def test_submit_csm_sar_ref_num_limit(self):
+        """
+        The SAR reference number is set correctly when the generated reference
+        number is larger than 0xFFFF.
+        """
+        service = yield self.get_service({'send_multipart_sar': True})
+        yield self.fake_smsc.bind()
+        # forward until we go past 0xFFFF
+        yield self.set_sequence_number(service, 0x10000)
+
+        long_message = 'This is a long message.' * 20
+        seq_nums = yield service.submit_csm_sar(
+            'abc123', 'dest_addr', short_message=long_message)
+        pdus = yield self.fake_smsc.await_pdus(4)
+        msg_parts = []
+        msg_refs = []
+
+        for i, sm in enumerate(pdus):
+            pdu_opts = unpacked_pdu_opts(sm)
+            mandatory_parameters = sm['body']['mandatory_parameters']
+
+            self.assertEqual('submit_sm', sm['header']['command_id'])
+            msg_parts.append(mandatory_parameters['short_message'])
+            self.assertTrue(len(mandatory_parameters['short_message']) <= 130)
+            msg_refs.append(pdu_opts['sar_msg_ref_num'])
+            self.assertEqual(i + 1, pdu_opts['sar_segment_seqnum'])
+            self.assertEqual(4, pdu_opts['sar_total_segments'])
+
+        self.assertEqual(long_message, ''.join(msg_parts))
+        self.assertEqual([2, 2, 2, 2], msg_refs)
+
+        stored_ids = yield self.lookup_message_ids(service, seq_nums)
+        self.assertEqual(['abc123'] * len(seq_nums), stored_ids)
+
+    @inlineCallbacks
+    def test_submit_csm_sar_single_part(self):
+        """
+        If the content fits in a single message, all the multipart madness is
+        avoided.
+        """
+        service = yield self.get_service({'send_multipart_sar': True})
+        yield self.fake_smsc.bind()
+
+        content = 'a' * 160
+        seq_numbers = yield service.submit_csm_sar(
+            'abc123', 'dest_addr', short_message=content)
+        self.assertEqual(len(seq_numbers), 1)
+        submit_sm_pdu = yield self.fake_smsc.await_pdu()
+
+        self.assertEqual(command_id(submit_sm_pdu), 'submit_sm')
+        self.assertEqual(short_message(submit_sm_pdu), content)
+        self.assertEqual(unpacked_pdu_opts(submit_sm_pdu), {})
+
+    @inlineCallbacks
+    def test_submit_csm_udh(self):
+        """
+        A long message can be sent in multiple PDUs with carefully handcrafted
+        user data headers.
+        """
+        service = yield self.get_service({'send_multipart_udh': True})
+        yield self.fake_smsc.bind()
+
+        long_message = 'This is a long message.' * 20
+        seq_numbers = yield service.submit_csm_udh(
+            'abc123', 'dest_addr', short_message=long_message)
+        pdus = yield self.fake_smsc.await_pdus(4)
+        self.assertEqual(len(seq_numbers), 4)
+
+        msg_parts = []
+        msg_refs = []
+
+        for i, sm in enumerate(pdus):
+            mandatory_parameters = sm['body']['mandatory_parameters']
+            self.assertEqual('submit_sm', sm['header']['command_id'])
+            msg = mandatory_parameters['short_message']
+
+            udh_hlen, udh_tag, udh_len, udh_ref, udh_tot, udh_seq = [
+                ord(octet) for octet in msg[:6]]
+            self.assertEqual(5, udh_hlen)
+            self.assertEqual(0, udh_tag)
+            self.assertEqual(3, udh_len)
+            msg_refs.append(udh_ref)
+            self.assertEqual(4, udh_tot)
+            self.assertEqual(i + 1, udh_seq)
+            self.assertTrue(len(msg) <= 136)
+            msg_parts.append(msg[6:])
+            self.assertEqual(0x40, mandatory_parameters['esm_class'])
+
+        self.assertEqual(long_message, ''.join(msg_parts))
+        self.assertEqual(1, len(set(msg_refs)))
+
+        stored_ids = yield self.lookup_message_ids(service, seq_numbers)
+        self.assertEqual(['abc123'] * len(seq_numbers), stored_ids)
+
+    @inlineCallbacks
+    def test_submit_csm_udh_ref_num_limit(self):
+        """
+        User data headers are crafted correctly when the generated reference
+        number is larger than 0xFF.
+        """
+        service = yield self.get_service({'send_multipart_udh': True})
+        yield self.fake_smsc.bind()
+        # forward until we go past 0xFF
+        yield self.set_sequence_number(service, 0x100)
+
+        long_message = 'This is a long message.' * 20
+        seq_numbers = yield service.submit_csm_udh(
+            'abc123', 'dest_addr', short_message=long_message)
+        pdus = yield self.fake_smsc.await_pdus(4)
+        self.assertEqual(len(seq_numbers), 4)
+
+        msg_parts = []
+        msg_refs = []
+
+        for i, sm in enumerate(pdus):
+            mandatory_parameters = sm['body']['mandatory_parameters']
+            self.assertEqual('submit_sm', sm['header']['command_id'])
+            msg = mandatory_parameters['short_message']
+
+            udh_hlen, udh_tag, udh_len, udh_ref, udh_tot, udh_seq = [
+                ord(octet) for octet in msg[:6]]
+            self.assertEqual(5, udh_hlen)
+            self.assertEqual(0, udh_tag)
+            self.assertEqual(3, udh_len)
+            msg_refs.append(udh_ref)
+            self.assertEqual(4, udh_tot)
+            self.assertEqual(i + 1, udh_seq)
+            self.assertTrue(len(msg) <= 136)
+            msg_parts.append(msg[6:])
+            self.assertEqual(0x40, mandatory_parameters['esm_class'])
+
+        self.assertEqual(long_message, ''.join(msg_parts))
+        self.assertEqual(1, len(set(msg_refs)))
+
+        stored_ids = yield self.lookup_message_ids(service, seq_numbers)
+        self.assertEqual(['abc123'] * len(seq_numbers), stored_ids)
+
+    @inlineCallbacks
+    def test_submit_csm_udh_single_part(self):
+        """
+        If the content fits in a single message, all the multipart madness is
+        avoided.
+        """
+        service = yield self.get_service({'send_multipart_udh': True})
+        yield self.fake_smsc.bind()
+
+        content = 'a' * 160
+        seq_numbers = yield service.submit_csm_udh(
+            'abc123', 'dest_addr', short_message=content)
+        self.assertEqual(len(seq_numbers), 1)
+        submit_sm_pdu = yield self.fake_smsc.await_pdu()
+
+        self.assertEqual(command_id(submit_sm_pdu), 'submit_sm')
+        self.assertEqual(short_message(submit_sm_pdu), content)
+        self.assertEqual(
+            submit_sm_pdu['body']['mandatory_parameters']['esm_class'], 0)
